[#45] run-tui 기능을 별도 모듈로 분리
